feat: Add 2 bookmark toggle functionality
- Save toggled bookmark value in local storage
- Remove unbookmarked card from bookmark page
